Redo the fix.

+++ src/track/track_api.c
@@ -89,8 +89,8 @@ static s32 adjust_tow_by_bit_fifo_delay(tracker_t *tracker,
                                         const nav_data_sync_t *to_tracker) {
   s32 TOW_ms = TOW_INVALID;
   /* Compute time since the pending data was read from the FIFO */
-  u8 fifo_length =
-      (u8)((tracker->nav_bit_fifo.write_index) - (to_tracker->read_index));
+  u16 fifo_length =
+      (u16)((tracker->nav_bit_fifo.write_index) - (to_tracker->read_index));
   u32 fifo_time_diff_ms = fifo_length * tracker->bit_sync.bit_length;
 
   /* Add full bit times + fractional bit time to the specified TOW */
